end-of-day 24 Feb 2019
Hajra has made much progress on the team section. However, she had issue with allocating an even amount of space for each officer. Jose suggests setting the maximum width and height of all planet images to be the standard dimensions for all planet images. This will be attempted in the morning.