Volunteer Page added; temporarily made the default page.
The volunteer page was added as a component.
We temporarily made it the default page in order to make testing easier. However, the Home page should be made the default page again later.
TODO: The body background color should change according to the page. We need to set it programmatically.