Unexpected deviation from tutorial
We're using the following tutorial: https://fireship.io/lessons/angular-universal-firebase/
According to it, we're supposed to get an error on step 3. We didn't (although a warning came up, which seemed to agree with the tutorial's concerns).
The tutorial would've recommended installing pollyfills had the issue occurred. We'll do that anyways.